Finish importing logic to file-loader-context
