+++ features/facebookImport/src/context/file-loader-context.jsx
@@ -1,37 +1,87 @@
 import React, { createContext, useContext, useEffect, useState } from "react";
-import { ImporterContext } from "./importer-context.jsx";
+import Storage from "../model/storage.js";
 import { analyzeFile } from "../model/analysis.js";
 import { importData } from "../model/importer.js";
 
+import { RefreshFilesError } from "../errors/polyIn-errors.js";
+
+//used until real storage is loaded
+const fakeStorage = {
+    files: null,
+    refreshFiles: async () => null,
+    readFile: async () => null,
+    removeFile: async () => {},
+};
+
 export const FileLoaderContext = createContext();
 
-export const FileLoaderProvider = ({ children }) => {
-    const { files } = useContext(ImporterContext);
+export const FileLoaderProvider = ({ children, parentContext }) => {
+    const { pod, setIsLoading, setGlobalError } = useContext(parentContext);
 
-    const [facebookAccount, setFacebookAccount] = useState(null);
+    const [storage, setStorage] = useState(fakeStorage);
+    const [files, setFiles] = useState(null);
+    const [account, setAccount] = useState(null);
     const [fileAnalysis, setFileAnalysis] = useState(null);
+
+    function refreshFiles() {
+        setIsLoading(true);
+        storage
+            .refreshFiles()
+            .then(async () => {
+                const resolvedFiles = [];
+                if (!storage.files) {
+                    setFiles(null);
+                    return;
+                }
+                for (const file of storage.files) {
+                    resolvedFiles.push(await file);
+                }
+                setFiles(resolvedFiles);
+                setIsLoading(false);
+            })
+            .catch((error) => setGlobalError(new RefreshFilesError(error)));
+    }
+
+    const handleRemoveFile = (fileID) => {
+        setAccount(null);
+        return storage.removeFile(fileID);
+    };
+
+    useEffect(() => {
+        if (!pod) return;
+        const storage = new Storage(pod);
+        storage.changeListener = async () => {
+            const resolvedFiles = [];
+            for (const file of storage.files) {
+                resolvedFiles.push(await file);
+            }
+            setFiles(Object.values(resolvedFiles));
+        };
+        setStorage(storage);
+    }, [pod]);
+
+    //on storage change
+    useEffect(() => {
+        refreshFiles();
+    }, [storage]);
 
     //on file change
     //when files changed run the importer first and create an account model first.
     //after there is an account the analyses are triggered.
     useEffect(() => {
-        if (files?.[0])
-            importData(files[0]).then((newFacebookAccount) =>
-                setFacebookAccount(newFacebookAccount)
-            );
-    }, [files]);
-
-    // On account changed
-    // When the account changes run the analises
-    useEffect(() => {
-        if (facebookAccount && files)
-            analyzeFile(files[0], facebookAccount).then((fileAnalysis) =>
+        if (!files?.[0]) return;
+        importData(files[0]).then((newAccount) => {
+            setAccount(newAccount);
+            analyzeFile(files[0], newAccount).then((fileAnalysis) =>
                 setFileAnalysis(fileAnalysis)
             );
-    }, [facebookAccount, files]);
+        });
+    }, [files]);
 
     return (
-        <FileLoaderContext.Provider value={{ fileAnalysis, facebookAccount }}>
+        <FileLoaderContext.Provider
+            value={{ files, fileAnalysis, account, handleRemoveFile }}
+        >
             {children}
         </FileLoaderContext.Provider>
     );
